+++ tests/Framebuffer Test/main.cpp
@@ -170,7 +170,7 @@ bool initSubsystems() {
     
     if (SDL_Init(SDL_INIT_EVERYTHING) < 0) {
         LS_LOG_GAME_ERR(
-            "Unable to initialize SDL due to error ", (int)SDL_GetError(), '\n',
+            "Unable to initialize SDL due to error ", SDL_GetError(), '\n',
             "Complain to your local programmer.\n"
         );
         return false;
